pyproject.toml
[project]
name = "semanticfilesearcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "faiss-cpu>=1.13.2",
    "google-generativeai>=0.8.6",
    "openai>=2.16.0",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "pyside6>=6.10.1",
    "python-docx>=1.2.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.0.0",
    "watchdog>=6.0.0",
]
